requirements.txt
gradio==5.9.1
openai==1.59.3
Pillow==10.3.0
requests==2.31.0
python-dotenv==1.0.1